+++ docs/source/changelog.rst
@@ -9,6 +9,8 @@ Release date: TBD
 
 New features
 ------------
+- Support for julia `Project.toml`, `JuliaProject.toml` and `Manifest.toml` files in :pr:`595` by
+  :user:`davidanthoff`
 
 API changes
 -----------